add recommend item (진행중)

+++ app/src/main/java/com/example/cosmeticlistproject/ui/view/MainActivity.kt
@@ -1,15 +1,16 @@
 package com.example.cosmeticlistproject.ui.view
 
 import android.os.Bundle
+import android.util.Log
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.cosmeticlistproject.R
 import com.example.cosmeticlistproject.base.BaseActivity
+import com.example.cosmeticlistproject.data.Recommend
 import com.example.cosmeticlistproject.databinding.ActivityMainBinding
 import com.example.cosmeticlistproject.ui.adapter.ProductListAdapter
 import com.example.cosmeticlistproject.ui.viewmodel.ProductViewModel
-import com.example.cosmeticlistproject.util.StateResult
 
 class MainActivity : BaseActivity<ActivityMainBinding, ProductViewModel>() {
     override val layoutResID: Int = R.layout.activity_main
@@ -18,6 +19,8 @@ class MainActivity : BaseActivity<ActivityMainBinding, ProductViewModel>() {
     val productListAdapter by lazy {
         ProductListAdapter(this)
     }
+
+    var recommendLists = ArrayList<ArrayList<Recommend>>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -30,9 +33,15 @@ class MainActivity : BaseActivity<ActivityMainBinding, ProductViewModel>() {
         viewDataBinding.rvProductList.layoutManager = LinearLayoutManager(applicationContext)
         viewDataBinding.rvProductList.adapter = productListAdapter
 
-        viewModel.getResult(currentPage)
+        viewModel.getProductResult(currentPage)
+        viewModel.getRecommendResult()
         viewModel.productResponse.observe(this, Observer { response ->
             productListAdapter.addProducts(response.productList)
+        })
+        viewModel.recommendResponse.observe(this, Observer { response ->
+            recommendLists.add(response.recommendList1)
+            recommendLists.add(response.recommendList2)
+            recommendLists.add(response.recommendList3)
         })
     }
 
@@ -41,12 +50,14 @@ class MainActivity : BaseActivity<ActivityMainBinding, ProductViewModel>() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
 
-                if(!viewDataBinding.rvProductList.canScrollVertically(1)) {
-                    productListAdapter.deleteLoading()
+                val lastVisibleItemPosition = (recyclerView.layoutManager as LinearLayoutManager?)!!.findLastCompletelyVisibleItemPosition()
+                val itemTotalCount = recyclerView.adapter!!.itemCount - 1
 
-                    if(viewModel.responseResult.value != StateResult.ERROR) {
-                        viewModel.getResult(++currentPage)
-                    }
+//                    productListAdapter.addRecommends(recommendLists[lastVisibleItemPosition / 11], lastVisibleItemPosition)
+
+                if(!viewDataBinding.rvProductList.canScrollVertically(1) && lastVisibleItemPosition == itemTotalCount) {
+                    productListAdapter.deleteLoading()
+                    viewModel.getProductResult(++currentPage)
                 }
             }
         })
